mv tests into conformance-tests package
