+++ packages/conformance-tests/src/lib.rs
@@ -1,0 +1,3 @@
+#![cfg(test)]
+
+mod resolver;
