+++ packages/conformance-tests/src/resolver.rs
@@ -1,0 +1,4 @@
+//! Recursive resolver role
+
+mod dns;
+mod dnssec;
