+++ packages/conformance-tests/src/resolver/dns.rs
@@ -1,0 +1,3 @@
+//! plain DNS functionality
+
+mod scenarios;
