+++ packages/conformance-tests/src/resolver/dnssec.rs
@@ -1,0 +1,3 @@
+//! DNSSEC functionality
+
+mod scenarios;
